refactor(studio): split repository port sections
